fix: use generic sans-serif for PlantUML SVG images
When using SVG via the <img> tag in HTML, fonts need to be included in the SVG.
(It's not possible to load fonts via external urls.)

Somehow it works for a local webserver, but not via GitHub pages.

Therefore: just use generic sans-serif for now.

Also: add a hint about showing PlantUML skinparameters.